Pass custom build-arg to Tekton 4.17 pipelines
The build was failing because the INPUT_DIR build argument was not being
passed to the buildah command.

The `extract-and-validate` step was failing because the build was using a
non-certified base image. This change updates the Tekton configurations
to use a certified Red Hat OPM image, which will resolve the
validation failure.

Signed-off-by: Daniel Farrell <dfarrell@redhat.com>